feat: community quality hardening
- Add pagination overflow guards + tests\n- Harden idempotency/internal-client behavior\n- Improve websocket presence/metrics\n- Tighten metrics basic-auth defaults\n- Add frontend idempotency key cache + tests

+++ backend/community-bootstrap/src/main/java/com/nowcoder/community/social/follow/RedisFollowRepository.java
@@ -124,7 +124,7 @@ public class RedisFollowRepository implements FollowRepository {
         Set<ZSetOperations.TypedTuple<String>> tuples = redisTemplate.opsForZSet().reverseRangeWithScores(
                 followeeKey(userId, entityType),
                 offset,
-                offset + limit - 1L
+                (long) offset + (long) limit - 1L
         );
         return toItems(tuples);
     }
@@ -134,7 +134,7 @@ public class RedisFollowRepository implements FollowRepository {
         Set<ZSetOperations.TypedTuple<String>> tuples = redisTemplate.opsForZSet().reverseRangeWithScores(
                 followerKey(entityType, entityId),
                 offset,
-                offset + limit - 1L
+                (long) offset + (long) limit - 1L
         );
         return toItems(tuples);
     }
